refactor: それぞれの処理のコードを分ける
