+++ functions/src/controller/PingController.ts
@@ -1,0 +1,5 @@
+import { Request, Response } from "firebase-functions"; 
+
+export async function PingController(req: Request, res: Response) {
+  res.json({msg: "Hello, firebase functions."});
+}
